feat: move cognito and api gateway from the backend stack to the static asset stack
