+++ cdk/bin/nishikiBackend.ts
@@ -15,4 +15,7 @@ const staticAssets = new NishikiStaticAssetsStack(
 );
 new NishikiBackendStack(app, `NishikiBackendStack-${stage}`, {
 	stage,
+	api: staticAssets.restApi,
+	userPool: staticAssets.userPool,
+	table: staticAssets.table
 });
