--- cdk/test/nishikiBackend-stack.test.ts
@@ -1,30 +1,0 @@
-import * as cdk from "aws-cdk-lib";
-import { Template } from "aws-cdk-lib/assertions";
-import { NishikiBackendStack } from "../lib/nishikiBackend-stack";
-
-describe("NishikiBackendStack", () => {
-	test("synthesizes correctly", () => {
-		const app = new cdk.App();
-
-		// Create the NishikiBackendStack.
-		const nishikiBackendStack = new NishikiBackendStack(
-			app,
-			"NishikiBackendStack",
-			{ stage: "dev" },
-		);
-
-		// Prepare the stack for assertions.
-		const template = Template.fromStack(nishikiBackendStack);
-
-		// Assert it creates the user pool with the correct Schema properties
-		template.hasResourceProperties("AWS::Cognito::UserPool", {
-			Schema: [{ Mutable: true, Name: "email", Required: true }],
-		});
-
-		// Assert it creates the user pool client with the correct SupportedIdentityProviders properties
-		// TODO: This should be passed.
-		// template.hasResourceProperties("AWS::Cognito::UserPoolClient", {
-		// 	SupportedIdentityProviders: ["Google"],
-		// });
-	});
-});
