+++ cdk/test/nishikiStaticAssets-stack.test.ts
@@ -94,6 +94,20 @@ describe("Static Assets", () => {
 			],
 		});
 	});
+
+	test("Cognito", () => {
+		// Assert it creates the user pool with the correct Schema properties
+		template.hasResourceProperties("AWS::Cognito::UserPool", {
+			UserPoolName: "nishiki-users-prod-user-pool",
+			Schema: [{ Mutable: true, Name: "email", Required: true }],
+		});
+
+		// Assert it creates the user pool client with the correct SupportedIdentityProviders properties
+		// TODO: This should be passed.
+		// template.hasResourceProperties("AWS::Cognito::UserPoolClient", {
+		// 	SupportedIdentityProviders: ["Google"],
+		// });
+	})
 });
 
 describe("dev environment", () => {
@@ -113,4 +127,10 @@ describe("dev environment", () => {
 			TableName: "nishiki-table-dev-db",
 		});
 	});
+
+	test("Cognito name", () => {
+		template.hasResourceProperties("AWS::Cognito::UserPool", {
+			UserPoolName: "nishiki-users-dev-user-pool",
+		});
+	})
 });
